replace small amount of flags with svg
the flag overlay should on its own improve perceived quality of the flags

this ideally should result in replacing of all of the flags, however this may not be feasible due to the large complexity of certain flags. i want to keep this from ballooning in size, even if it is moved externally, so modifying flags to reduce file size might be necessary.

there are over 190 flags left to do